% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbar.R
\name{badr}
\alias{badr}
\title{Detect anomaly}
\usage{
badr(.data, pre.period, post.period, verbose = getOption("badr.verbose"), ...)
}
\arguments{
\item{.data}{data table}

\item{pre.period}{vector}

\item{post.period}{vector}

\item{...}{params for \code{bsts_model}}
}
\description{
Detect anomaly
}
