% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsts.R
\name{bsts_model}
\alias{bsts_model}
\title{Create bsts model}
\usage{
bsts_model(.data, spec = NULL, ...)
}
\arguments{
\item{.data}{training set}

\item{spec}{\code{cbar.model.spec} object}

\item{...}{params for \code{bsts_spec_static}}
}
\value{
\code{bsts} which is a bsts model
}
\description{
Create bsts model
}
\references{
Scott, S. L., & Varian, H. R. (2014). Predicting the present with
  bayesian structural time series. International Journal of Mathematical
  Modelling and Numerical Optimisation, 5(1-2), 4-23.
}
