% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsts.R
\name{bsts_spec_static}
\alias{bsts_spec_static}
\title{Specify bsts model for static linear regression}
\usage{
bsts_spec_static(.data, sigma_guess = NULL, upper_limit = NULL,
  sd_prior_sample_size = 32, expected_model_size = 3,
  expected_r2 = 0.8, prior_df = 50, niter = 1000, ping = 0,
  model_options = NULL, nseasons = 0, ...)
}
\arguments{
\item{.data}{time-series data to be trained}

\item{sigma_guess}{an argument for \code{bsts::bsts}}

\item{upper_limit}{an argument for \code{bsts::bsts}}

\item{sd_prior_sample_size}{an argument for \code{bsts::bsts}}

\item{expected_model_size}{an argument for \code{bsts::bsts}}

\item{expected_r2}{an argument for \code{bsts::bsts}}

\item{prior_df}{an argument for \code{bsts::bsts}}

\item{niter}{an argument for \code{bsts::bsts}}

\item{ping}{an argument for \code{bsts::bsts}}

\item{model_options}{an argument for \code{bsts::bsts}}

\item{nseasons}{number of seasons}

\item{...}{params for \code{bsts_model}}
}
\value{
\code{cbar.model.spec} object for model specification
}
\description{
Specify bsts model for static linear regression
}
