% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbar.R
\name{cbar}
\alias{cbar}
\title{Detect contextual anomaly}
\usage{
cbar(.data, pre.period, post.period, verbose = getOption("badr.verbose"), ...)
}
\arguments{
\item{.data}{data table}

\item{pre.period}{vector}

\item{post.period}{vector}

\item{...}{params for \code{bsts_model}}
}
\description{
Detect contextual anomaly
}
