% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{destandardized}
\alias{destandardized}
\title{Destandardize a vector}
\usage{
destandardized(y_hat, y_mu, y_sd)
}
\arguments{
\item{y_hat}{standardized numeric vector}

\item{y_mu}{a mean value of unstandarized vector}

\item{y_sd}{a standard deviation of unstandarized vector}
}
\description{
Destandardize a vector
}
