% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{destandarized}
\alias{destandarized}
\title{Destandarized a vector}
\usage{
destandarized(y_hat, y_mu, y_sd)
}
\arguments{
\item{y_hat}{standarized numeric vector}

\item{y_mu}{a mean value of unstandarized vector}

\item{y_sd}{a standard deviation of unstandarized vector}
}
\description{
Destandarized a vector
}
