% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abs-visual.R
\name{plot_error}
\alias{plot_error}
\title{Print estimation error plot}
\usage{
plot_error(.cbar, xlab = "", ylab = "Estimation error", method = "diff",
  ...)
}
\arguments{
\item{.cbar}{cbar object}

\item{method}{diff}

\item{x_label}{a label for x-axis}

\item{y_label}{a label for y-ayis}
}
\value{
\code{boxplot} object
}
\description{
Print estimation error plot
}
