% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abs-visual.R
\name{plot_error_}
\alias{plot_error_}
\title{Print estimation error plot}
\usage{
plot_error_(.error, xlab = "", ylab = "Estimation error",
  method = "diff", ...)
}
\arguments{
\item{.error}{error data frame}

\item{xlab}{a label for x-axis}

\item{ylab}{a label for y-ayis}

\item{method}{diff}

\item{...}{params for boxplot}
}
\description{
Print estimation error plot
}
