% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abs-visual.R
\name{plot_incprob}
\alias{plot_incprob}
\title{Print inclusion probablity plot}
\usage{
plot_incprob(.cbar, threshold = 0.1, horiz = T, cex.names = 0.5,
  xlab = "Inclusion probability (\%)", las = 1, ...)
}
\arguments{
\item{.cbar}{cbar object}

\item{threshold}{a threhold for inclusion probablity}
}
\value{
\code{boxplot} object
}
\description{
Print inclusion probablity plot
}
