% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abs-visual.R
\name{plot_incprob_}
\alias{plot_incprob_}
\title{Print inclusion probablity plot}
\usage{
plot_incprob_(.incprob, threshold = 0.1, horiz = T, cex.names = 0.5,
  xlab = "Inclusion probability (\%)", las = 1, ...)
}
\arguments{
\item{.incprob}{data frame}

\item{threshold}{a threhold for inclusion probablity}

\item{horiz}{horiz}

\item{cex.names}{cex.names}

\item{xlab}{xlab}

\item{las}{las}

\item{...}{params for barplot}
}
\value{
\code{boxplot} object
}
\description{
Print inclusion probablity plot
}
