% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abs_visual.R
\name{plot_ts}
\alias{plot_ts}
\title{Visual abstraction with time series plot}
\usage{
plot_ts(.cbar, x_label = "", y_label = "", seq_by = NULL)
}
\arguments{
\item{.cbar}{cbar object}
}
\value{
\code{ggplot} object
}
\description{
Visual abstraction with time series plot
}
