% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{point_prediction}
\alias{point_prediction}
\title{Get point prediction from posterior means and response trajectories}
\usage{
point_prediction(y_hat, .posterior_mean, alpha = 0.05)
}
\arguments{
\item{alpha}{alpha}

\item{yhat}{response trajectories}

\item{.posetrior_mean}{posterior mean values}
}
\description{
Get point prediction from posterior means and response trajectories
}
