% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{posterior_mean}
\alias{posterior_mean}
\title{Generate posterior mean of the response variable}
\usage{
posterior_mean(.model)
}
\arguments{
\item{.model}{bsts model}
}
\description{
This one is used for point prediction based one predictive posterior
distribution
}
