% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbar.R
\name{print.cbar}
\alias{print.cbar}
\title{Print cbar object}
\usage{
\method{print}{cbar}(x, ...)
}
\arguments{
\item{x}{\code{cbar} object to print}

\item{...}{further arguments passed to or from other methods}
}
\description{
Print cbar object
}
