% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abs-analytical.R
\name{summarise_anomaly}
\alias{summarise_anomaly}
\title{Summarise anomaly detection result}
\usage{
summarise_anomaly(.cbar, .session = NULL)
}
\arguments{
\item{.cbar}{\code{cbar} object}

\item{.session}{names of sessions, which is NULL by default}
}
\value{
\code{data.frame} that summarises input data with anomaly label
}
\description{
Summarise anomaly detection result
}
