% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abs-analytical.R
\name{summarise_incprob}
\alias{summarise_incprob}
\title{Summarise inclusion probability of model}
\usage{
summarise_incprob(.cbar, threshold = 0.1)
}
\arguments{
\item{.cbar}{\code{cbar} object}

\item{threshold}{threshold of inclusion probability, which is .1 by default}
}
\value{
\code{vector} that summarises inclusion probabilities for each MCMC
  samples
}
\description{
Summarise inclusion probability of model
}
