% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abs-analytical.R
\name{summarise_session}
\alias{summarise_session}
\title{Summarise anomaly in session}
\usage{
summarise_session(.cbar, .session = NULL)
}
\arguments{
\item{.cbar}{\code{cbar} object}

\item{.session}{names of sessions, which is NULL by default}
}
\value{
\code{data.frame} that summarises outcome for each sessoin
}
\description{
Summarise anomaly in session
}
