% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsts.R
\name{bsts_model}
\alias{bsts_model}
\title{Create bsts model}
\usage{
bsts_model(.data, spec = NULL)
}
\arguments{
\item{.data}{training set}

\item{spec}{bsts model specification}
}
\description{
Create bsts model
}
