% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbar.R
\name{cbar}
\alias{cbar}
\title{Detect contextual anomaly}
\usage{
cbar(.data, ref_period, mea_period, apply_standardized = T,
  verbose = getOption("cbar.verbose"), ...)
}
\arguments{
\item{.data}{data table with datetime, y, and predictors}

\item{ref_period}{performance reference period}

\item{mea_period}{performance measurement period}

\item{...}{params for \code{bsts_model}}
}
\description{
In anomaly detection, especailly in telecommunication, we call it reference
and measurement period
}
