% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbar.R
\name{cbar}
\alias{cbar}
\title{Detect contextual anomaly}
\usage{
cbar(.data, pre_period, post_period, verbose = getOption("cbar.verbose"), ...)
}
\arguments{
\item{.data}{data table}

\item{pre_period}{vector}

\item{post_period}{vector}

\item{...}{params for \code{bsts_model}}
}
\description{
Detect contextual anomaly
}
