% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{inference}
\alias{inference}
\title{Get inference from predictive posetrior prediction of bsts model}
\usage{
inference(.model, y, post_period, alpha = 0.05)
}
\description{
Get inference from predictive posetrior prediction of bsts model
}
