% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abs-visual.R
\name{plot_ts}
\alias{plot_ts}
\title{Print time-series plot}
\usage{
plot_ts(.cbar, x_label = "", y_label = "", seq_by = NULL)
}
\arguments{
\item{.cbar}{cbar object}

\item{x_label}{a label for x-axis}

\item{y_label}{a label for y-ayis}

\item{seq_by}{increment of the sequence, which is NULL by default}
}
\value{
\code{ggplot} object
}
\description{
Print time-series plot
}
