% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{response_trajectory}
\alias{response_trajectory}
\title{Generate trajectories of the response variable}
\usage{
response_trajectory(.model, seed = NULL)
}
\arguments{
\item{.model}{\code{bsts_model}}

\item{seed}{a seed value. NULL by default}
}
\value{
\code{data.frame} that contains predicted value \code{y_hat}
}
\description{
Note that \code{posterior_state_samples} returns posterior mean, whereas,
this one returns posterior mean + noise
}
\details{
This one is used for lower and upper bounds
}
